Convert Services page components from Next.js to React Router
